successful restructuring of tests file. Changes to be committed: 	modified:   tests/__init__.py 	deleted:    tests/test_answer.py 	new file:   tests/test_auth.py 	new file:   tests/test_question.py 	deleted:    tests/test_questions.py 	renamed:    tests/test_endpoints.py -> tests/test_questions_answer.py 	deleted:    tests/test_user.py
