+++ tests/test_questions_answer.py
@@ -1,43 +1,11 @@
-import pytest
+from app.models.answer import Answer
 from tests import (client, post_json, post_json_header,
  signin, user_two, put_json_header)
+import pytest
 
-def test_user_creation(client):
-    resp = post_json(client, '/v1/auth/signup', { 
-	"email": "test@test.com",
-    "name": "test",
-	"password":"testAs1v"})
-    assert b'User successfully created' in resp.data
-    assert resp.status_code == 201
-
-def test_duplicate_user_creation(client):
-    resp = post_json(client, '/v1/auth/signup', { 
-	"email": "test@test.com",
-    "name": "test",
-	"password":"testAs1v"})
-    assert b'Email address already exists' in resp.data
-    assert resp.status_code == 400
-
-def test_user_login(client):
-    resp = post_json(client, '/v1/auth/signin', { 
-	"email": "test@test.com",
-	"password":"testAs1v"})
-    assert b'Successful login' in resp.data
-    assert resp.status_code == 200
-
-def test_add_question(client):
-    resp = post_json_header(client, '/v1/questions', {
-        "title": "big is big",}, 
-    headers={'Authorization': 'Bearer ' + signin(client)})
-    assert resp.status_code == 201
-
-def test_get_questiosn(client):
-    resp = client.get('/v1/questions', headers={'Authorization': 'Bearer ' + signin(client)})
-    assert resp.status_code == 200
-
-def test_get_single_question(client):
-    resp = client.get('/v1/questions/1', headers={'Authorization': 'Bearer ' + signin(client)})
-    assert b'title' in resp.data
+def test_is_instance_of_answer():
+    new_answer = Answer(5, "Hello there", "Nangai")
+    assert isinstance(new_answer, Answer)
 
 def test_post_answer(client):
     resp = post_json_header(client,'/v1/questions/1/answers', {
@@ -77,5 +45,3 @@ def test_delete_question(client):
     resp = client.delete('/v1/questions/1/delete', headers={'Authorization': 'Bearer ' + signin(client)})
     assert resp.status_code == 200
     assert b'Question Deleted' in resp.data
-
-
